backend/pyproject.toml
[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "objectdms"
version = "1.0.0"
description = "Object Document Management System"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "beanie>=1.25.0",
    "motor>=3.3.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.0.0",
    "prometheus-client>=0.17.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.setuptools]
packages = ["app"] 